refactor(buildSchema): extract configType conditional logic from flow and typescript folders
